\name{apcluster-deprecated}
\alias{apcluster-deprecated}
\alias{sparseToFull}
\title{Deprecated functions in package \sQuote{apcluster}}
\description{
  These functions are provided for compatibility with older versions
  of \sQuote{apcluster} only, and will be defunct at the next release.
}
\details{
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{sparseToFull: \code{\link{as.DenseSimilarityMatrix}}}}
}
\author{Ulrich Bodenhofer & Andreas Kothmeier
\email{apcluster@bioinf.jku.at}}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{cluster}

